backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
pydantic
python-jose[cryptography]
bcrypt==4.0.1
sqlalchemy
psycopg2-binary
